requirements.txt
pyusb == 1.2.1
pillow == 10.4.0
